refactor: PopUp -> PhoneNumberPopUp
